requirements.txt
Django==1.11.15
django-ckeditor==5.6.1
django-extensions==2.1.0
django-js-asset==1.1.0
django-rest-framework==0.1.0
djangorestframework==3.8.2
Pillow==5.2.0
pytz==2018.5
six==1.11.0